Fix data labels in ggplotly-vs-plotly.Rmd
The data labels are not aligning to their respective segments. My guess is that this broke with ggplot2 2.2.0 which reversed the order stacking from the group order.

I've tested the change to ensure the results are correct with the most up-to-date version of plotly and ggplot2, but I am leaving any site-building changes to you as I do not know how that works and assume you have your way of doing that.